# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"ServiceResponse_any-Array_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"items": {},
						"type": "array",
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ServiceResponse_any-or-null_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ServiceResponse_null_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"anyOf": [],
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"VisitorCreateAttributes": {
				"properties": {
					"fullName": {
						"type": "string"
					},
					"mobile": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"visitorCompany": {
						"type": "string"
					},
					"purpose": {
						"type": "string"
					},
					"department": {
						"type": "string"
					},
					"hostName": {
						"type": "string"
					},
					"idProofType": {
						"type": "string"
					},
					"idNumber": {
						"type": "string"
					},
					"passType": {
						"type": "string"
					},
					"status": {
						"type": "string",
						"enum": [
							"IN",
							"OUT",
							"ACTIVE",
							"CHECKED_IN",
							"CHECKED_OUT",
							"BLACKLISTED"
						]
					},
					"badgeId": {
						"type": "string"
					},
					"profilePhoto": {
						"type": "string"
					},
					"entryTime": {
						"type": "string",
						"format": "date-time"
					},
					"exitTime": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"fullName"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PublicCheckInRequest": {
				"properties": {
					"visitor": {
						"properties": {
							"profilePhoto": {
								"type": "string"
							},
							"idNumber": {
								"type": "string"
							},
							"idProofType": {
								"type": "string"
							},
							"hostName": {
								"type": "string"
							},
							"department": {
								"type": "string"
							},
							"purpose": {
								"type": "string"
							},
							"visitorCompany": {
								"type": "string"
							},
							"email": {
								"type": "string"
							},
							"mobile": {
								"type": "string"
							},
							"fullName": {
								"type": "string"
							}
						},
						"required": [
							"fullName"
						],
						"type": "object"
					},
					"verificationMode": {
						"type": "string"
					},
					"department": {
						"type": "string"
					},
					"badgeType": {
						"type": "string",
						"enum": [
							"QR",
							"RFID"
						]
					},
					"visitorType": {
						"type": "string"
					},
					"date": {
						"type": "string"
					},
					"time": {
						"type": "string"
					}
				},
				"required": [
					"visitor"
				],
				"type": "object",
				"additionalProperties": {}
			},
			"ServiceResponse_any_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CheckInRequest": {
				"properties": {
					"visitor": {
						"properties": {
							"idNumber": {
								"type": "string"
							},
							"idProofType": {
								"type": "string"
							},
							"hostName": {
								"type": "string"
							},
							"department": {
								"type": "string"
							},
							"purpose": {
								"type": "string"
							},
							"visitorCompany": {
								"type": "string"
							},
							"email": {
								"type": "string"
							},
							"mobile": {
								"type": "string"
							},
							"fullName": {
								"type": "string"
							}
						},
						"required": [
							"fullName"
						],
						"type": "object"
					},
					"verificationMode": {
						"type": "string"
					},
					"appointmentId": {
						"type": "string"
					},
					"department": {
						"type": "string"
					},
					"badgeType": {
						"type": "string",
						"enum": [
							"QR",
							"RFID"
						]
					}
				},
				"required": [
					"visitor"
				],
				"type": "object",
				"additionalProperties": {}
			},
			"VisitorUpdateAttributes": {
				"properties": {
					"fullName": {
						"type": "string"
					},
					"mobile": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"visitorCompany": {
						"type": "string"
					},
					"purpose": {
						"type": "string"
					},
					"department": {
						"type": "string"
					},
					"hostName": {
						"type": "string"
					},
					"idProofType": {
						"type": "string"
					},
					"idNumber": {
						"type": "string"
					},
					"passType": {
						"type": "string"
					},
					"status": {
						"type": "string",
						"enum": [
							"IN",
							"OUT",
							"ACTIVE",
							"CHECKED_IN",
							"CHECKED_OUT",
							"BLACKLISTED"
						]
					},
					"badgeId": {
						"type": "string"
					},
					"profilePhoto": {
						"type": "string"
					},
					"entryTime": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"exitTime": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"AttendanceResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"checkIn": {
						"type": "string",
						"format": "date-time"
					},
					"checkOut": {
						"type": "string",
						"format": "date-time"
					},
					"hoster": {
						"type": "string"
					},
					"badge": {
						"type": "string"
					},
					"date": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"note": {
						"type": "string"
					},
					"user": {},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"userId",
					"checkIn",
					"date",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"AttendanceSingleResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"allOf": [
							{
								"$ref": "#/components/schemas/AttendanceResponse"
							}
						],
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CheckOutRequest": {
				"properties": {
					"note": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header",
				"description": "JWT token (Bearer <token>)"
			}
		}
	},
	"info": {
		"title": "Attendance Backend API",
		"version": "1.0.0",
		"description": "API documentation for attendance management system",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "API Support"
		}
	},
	"paths": {
		"/api/visitors": {
			"get": {
				"operationId": "GetAllVisitors",
				"responses": {
					"200": {
						"description": "Visitors retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Visitors"
				],
				"security": [
					{
						"jwt": [
							"visitor:read"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "query",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "department",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateVisitor",
				"responses": {
					"201": {
						"description": "Visitor created successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any-or-null_"
								}
							}
						}
					},
					"400": {
						"description": "Visitor fullName is required",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Visitors"
				],
				"security": [
					{
						"jwt": [
							"visitor:create"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/VisitorCreateAttributes"
							}
						}
					}
				}
			}
		},
		"/api/visitors/recent-taps": {
			"get": {
				"operationId": "GetRecentTaps",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Visitors"
				],
				"security": [
					{
						"jwt": [
							"visitor:read"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "department",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "query",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "searchType",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"name",
								"phone",
								"voice"
							]
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/visitors/search": {
			"get": {
				"operationId": "SearchVisitors",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Visitors"
				],
				"security": [
					{
						"jwt": [
							"visitor:read"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "query",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "searchType",
						"required": false,
						"schema": {
							"default": "name",
							"type": "string",
							"enum": [
								"name",
								"phone",
								"voice"
							]
						}
					},
					{
						"in": "query",
						"name": "department",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/visitors/public-check-in": {
			"post": {
				"operationId": "PublicCheckIn",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any_"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any_"
								}
							}
						}
					},
					"403": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"Visitors"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PublicCheckInRequest"
							}
						}
					}
				}
			}
		},
		"/api/visitors/check-in": {
			"post": {
				"operationId": "CheckIn",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any_"
								}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"Visitors"
				],
				"security": [
					{
						"jwt": [
							"visitor:create"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CheckInRequest"
							}
						}
					}
				}
			}
		},
		"/api/visitors/check-out": {
			"post": {
				"operationId": "CheckOut",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any_"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"Visitors"
				],
				"security": [
					{
						"jwt": [
							"visitor:update"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"badgeId": {
										"type": "string"
									},
									"visitorId": {
										"type": "string"
									}
								},
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/visitors/{visitorId}": {
			"get": {
				"operationId": "GetVisitorById",
				"responses": {
					"200": {
						"description": "Visitor retrieved successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any-or-null_"
								}
							}
						}
					},
					"404": {
						"description": "Visitor not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Visitors"
				],
				"security": [
					{
						"jwt": [
							"visitor:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "visitorId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateVisitor",
				"responses": {
					"200": {
						"description": "Visitor updated successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_any-or-null_"
								}
							}
						}
					},
					"404": {
						"description": "Visitor not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Visitors"
				],
				"security": [
					{
						"jwt": [
							"visitor:update"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "visitorId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/VisitorUpdateAttributes"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteVisitor",
				"responses": {
					"200": {
						"description": "Visitor deleted successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_null_"
								}
							}
						}
					},
					"404": {
						"description": "Visitor not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Visitors"
				],
				"security": [
					{
						"jwt": [
							"visitor:delete"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "visitorId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/attendance/{id}/checkout": {
			"put": {
				"operationId": "CheckOut",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AttendanceSingleResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Attendance"
				],
				"security": [
					{
						"jwt": [
							"attendance:update"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CheckOutRequest"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}